mv test into the corresponding class
